hazelcast_counter/requirements.txt
hazelcast-python-client==5.4.0